requirements.txt
beautifulsoup4
chromadb
faiss-cpu
flask
playwright
pypickle
python-dotenv
sentence_transformers
